Convert app to component, added state for it
