+++ src/components/app/app.jsx
@@ -1,15 +1,31 @@
-import React from "react";
+import React, {PureComponent} from "react";
 import Main from "../main/main";
 import PropTypes from "prop-types";
 
-const App = (props) => {
-  const clickHandler = () => {};
+class App extends PureComponent {
+  constructor(props) {
+    super(props);
 
-  return <Main
-    films = {props.films}
-    onclick={clickHandler}
-  />;
-};
+    this.state = {
+      activeCard: -1,
+    };
+  }
+
+  render() {
+    const clickHandler = () => {};
+
+    return <Main
+      films = {this.props.films}
+      onclick = {clickHandler}
+      onCardOver = {(index) => {
+        this.setState({
+          activeCard: parseInt(index, 10),
+        });
+        console.log(index);
+      }}
+    />;
+  }
+}
 
 App.propTypes = {
   films: PropTypes.array.isRequired,
