+++ src/components/app/app.test.js
@@ -2,9 +2,10 @@ import React from "react";
 import renderer from "react-test-renderer";
 import App from "./app";
 
+import {FILMS} from "./mocks/films";
 
 it(`Testing someone`, () => {
-  const FILMS = [`Fantastic Beasts`, `Bohemian Rhapsody`, `Macbeth`, `Aviator`];
+
   const tree = renderer.create(
       <App
         films = {FILMS}
