+++ src/components/films-list/films-list.jsx
@@ -2,17 +2,16 @@ import React from "react";
 import PropTypes from "prop-types";
 import FilmCard from "../film-card/film-card";
 
-
 const FilmsList = (props) => {
   let filmsList = [];
+  let index = 0;
 
   for (let film of props.films) {
     if (Object.keys(film).length !== 0) {
-      let index = 0;
-
       filmsList.push(<FilmCard
         film = {film}
         index = {index++}
+        onCardOver = {props.onCardOver}
       />);
     }
   }
@@ -22,6 +21,7 @@ const FilmsList = (props) => {
 
 FilmsList.propTypes = {
   films: PropTypes.array.isRequired,
+  onCardOver: PropTypes.func,
 };
 
 export default FilmsList;
