+++ src/mocks/films.js
@@ -1,27 +1,29 @@
-// const FILMS = [
-//   {
-//     title: `Fantastic Beasts: The Crimes of Grindelwald`,
-//     img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-//     src: `fantastic-beasts-the-crimes-of-grindelwald`,
-//   },
-//   {
-//     title: `Bohemian Rhapsody`,
-//     img: `img/bohemian-rhapsody.jpg`,
-//     src: `bohemian-rhapsody`,
-//   },
-//   {
-//     title: `Macbeth`,
-//     img: `img/macbeth.jpg`,
-//     src: `macbeth`,
-//   },
-//   {
-//     title: `Aviator`,
-//     img: `img/aviator.jpg`,
-//     src: `aviator`,
-//   },
-//   {
-//     title: `We need to talk about Kevin`,
-//     img: `img/we-need-to-talk-about-kevin.jpg`,
-//     src: `we-need-to-talk-about-kevin`,
-//   },
-// ];
+const FILMS = [
+  {
+    title: `Fantastic Beasts: The Crimes of Grindelwald`,
+    img: `img/fantastic-beasts-the-crimes-of-grindelwald.jpg`,
+    src: `fantastic-beasts-the-crimes-of-grindelwald`,
+  },
+  {
+    title: `Bohemian Rhapsody`,
+    img: `img/bohemian-rhapsody.jpg`,
+    src: `bohemian-rhapsody`,
+  },
+  {
+    title: `Macbeth`,
+    img: `img/macbeth.jpg`,
+    src: `macbeth`,
+  },
+  {
+    title: `Aviator`,
+    img: `img/aviator.jpg`,
+    src: `aviator`,
+  },
+  {
+    title: `We need to talk about Kevin`,
+    img: `img/we-need-to-talk-about-kevin.jpg`,
+    src: `we-need-to-talk-about-kevin`,
+  },
+];
+
+export {FILMS};
